Move extensions to separate namespace
