+++ Source/MochaTool.InteropGen/Extensions/CXCursorExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using ClangSharp.Interop;
 
-namespace MochaTool.InteropGen;
+namespace MochaTool.InteropGen.Extensions;
 
 /// <summary>
 /// Contains extension methods for the <see cref="CXCursor"/>.
